http/common: move common quic utils from http to quic
